Gemini func can now be accessed thru CLI, havent tried with Wails-frontend
