+++ backend/provider.go
@@ -1,17 +1,5 @@
 package main
 
-import "fmt"
-
-func main() {
-
-	var prompt string
-
-	fmt.Println("Please enter your prompt...")
-	fmt.Scanln(&prompt)
-	result := chatGPTProvider(prompt)
-	fmt.Println(result)
-}
-
 func check_online() {
 
 }
