+++ go.mod
@@ -1,0 +1,3 @@
+module monochrome-switcher
+
+go 1.25.0
